2006-SCSI-30/TripTally/backend/app/services/trafficcams/requirements.txt
# Traffic Camera CI Service Requirements

# Core
redis==7.0.1
numpy==2.2.6
pandas==2.3.3
requests==2.32.5

# Image Processing
opencv-python==4.12.0.88
Pillow==12.0.0

# ML Inference - YOLO
ultralytics>=8.0.0

# ML Training & Forecasting
scikit-learn>=1.5.0
xgboost>=2.0.0
joblib>=1.4.0

# Optional: for better performance
# If using ONNX instead of PyTorch:
# onnxruntime==1.23.2
# onnxruntime-gpu==1.23.2  # If you have CUDA GPU
